% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{draw_map}
\alias{draw_map}
\title{Map of locations with colours and/or sizes}
\usage{
draw_map(locations, colours = NULL, size = NULL, bounding_box,
  zoom = 5)
}
\arguments{
\item{locations}{data frame with columns Location, Longitude, Latitude}

\item{colours}{vector of things to colour, in same order as locations}

\item{bounding_box}{bottom left long, bottom left lat, top right long, top right lat (vector of length 4); W longs negative}

\item{zoom}{stamen maps zoom (higher=better resolution but more tiles to get), default 5}
}
\value{
map showing points labelled by location
}
\description{
Map of locations with colours and/or sizes
}
\examples{
draw_map(nine_points_locations, bounding_box=c(-87.5, 71, -82.5, 74), zoom=5)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
