% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{k_means}
\alias{k_means}
\title{Non-hierarchical (K-means) cluster analysis}
\usage{
k_means(d, n_cluster)
}
\arguments{
\item{n_cluster}{number of clusters to obtain}

\item{data}{frame in wide format (see make_wide)}
}
\value{
list: scree plot, cluster memberships
}
\description{
Non-hierarchical (K-means) cluster analysis
}
\examples{

k_means(nine_points, 4)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
