% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-maps.R
\name{make_everything}
\alias{make_everything}
\title{Make the whole analysis for a data set}
\usage{
make_everything(d, locations, bounding_box = F, zoom = 5,
  scaling = 1, n_cluster)
}
\arguments{
\item{d}{wide-format data set (see make_wide)}

\item{locations}{lats and longs of the locations (data frame with three columns Location, Longitude, Latitude)}

\item{bounding_box}{bounding box for map: bottom left long, lat; top right long, lat, longitudes W negative. Defaults to FALSE; then determined from data.}

\item{zoom}{zoom for map; higher requires more map tiles, but produces better map. Default 5, max 18.}

\item{scaling}{global scaling factor for points on Theil-Sen map, default 1 (bigger makes all points bigger)}

\item{n_cluster}{number of clusters to use for Ward and K-means maps}
}
\value{
list of 2: the temporal results from nake_temporal; the spatial results from make_spatial.
}
\description{
Make the whole analysis for a data set
}
\examples{

make_everything(nine_points, nine_points_locations, n_cluster=4)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
