% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_long.R
\name{make_long}
\alias{make_long}
\title{Make wide data long}
\usage{
make_long(d)
}
\arguments{
\item{d}{data frame in wide format (years in column 1, called `year`, locations in other columns)}
}
\value{
data frame in long format with one row per observation
}
\description{
Make wide data long
}
\examples{
make_long(nine_points)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
