% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-maps.R
\name{make_spatial}
\alias{make_spatial}
\title{Make all the maps for a data set}
\usage{
make_spatial(d, locations, bounding_box = F, zoom = 5, scaling = 1,
  n_cluster)
}
\arguments{
\item{d}{wide-format data set (see make_wide)}

\item{locations}{lats and longs of the locations (data frame with three columns Location, Longitude, Latitude)}

\item{bounding_box}{bounding box for map: bottom left long, lat; top right long, lat, longitudes W negative}

\item{zoom}{zoom for map; higher requires more map tiles, but produces better map. Default 5.}

\item{scaling}{global scaling factor for points on Theil-Sen map, default 1}

\item{n_cluster}{number of clusters to use for Ward and K-means maps}
}
\value{
list of maps (including scree plot and dendrograms)
}
\description{
Make all the maps for a data set
}
\examples{

make_spatial(nine_points, nine_points_locations, n_cluster=4)
make_spatial(nine_points, nine_points_locations, bounding_box=c(-87.5, 71, -82.5, 74), zoom=6, scaling=1, n_cluster=4)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
