% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-maps.R
\name{make_temporal}
\alias{make_temporal}
\title{Make temporal analysis without maps (in first part of spatial analysis)}
\usage{
make_temporal(d)
}
\arguments{
\item{d}{wide-format data set}
}
\value{
list: time trend plots with lowess curves and with linear trends, Mann-Kendall result table, summary of significance, Theil-Sen slopes table, summary.
}
\description{
Make temporal analysis without maps (in first part of spatial analysis)
}
\examples{

make_temporal(nine_points)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
