% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_long.R
\name{make_wide}
\alias{make_wide}
\title{Make long data wide}
\usage{
make_wide(d)
}
\arguments{
\item{d}{data frame in long format (year, location, day)}
}
\value{
data frame in wide format (column of years, one column per location)
}
\description{
Make long data wide
}
\examples{
make_wide(nine_points_long) # should be same as nine_points

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
