% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{map_cluster}
\alias{map_cluster}
\title{Map clusters from cluster analysis}
\usage{
map_cluster(clusters, locations, bounding_box, zoom, title)
}
\arguments{
\item{clusters}{cluster output from cluster analysis}

\item{locations}{data frame of lats and longs for locations (columns Location, Latitude, Longitude)}

\item{bounding_box}{(long and lat of bottom left, long and lat of top right)}

\item{zoom}{for stamen map (defaults to 5)}
}
\value{
map of locations coloured by cluster
}
\description{
Map clusters from cluster analysis
}
\examples{

map_cluster(ward(nine_points, T, 3)$clusters, nine_points_locations, bounding_box=c(-87.5, 71, -82.5, 74), zoom=6, title="Cluster map (Ward's method)")
map_cluster(k_means(nine_points, 4)$clusters, nine_points_locations, bounding_box=c(-87.5, 71, -82.5, 74), zoom=6, title="Cluster map (K-means method)")

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
