% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann-kendall.R
\name{mk_map}
\alias{mk_map}
\title{Map of Mann-Kendall significance levels}
\usage{
mk_map(d_long, locations, bounding_box, zoom = 5)
}
\arguments{
\item{d_long}{long format data set}

\item{locations}{lats and longs of locations}

\item{bounding_box}{bottom left long, lat, top right long, lat (as a vector)}

\item{zoom}{zoom factor for stamen map (default 5)}
}
\description{
Map of Mann-Kendall significance levels
}
\examples{
mk_map(nine_points_long, nine_points_locations, c(-87.5, 71, -82.5, 74), zoom=6)
}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
