% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann-kendall.R
\name{mk_sig}
\alias{mk_sig}
\title{Count number of significant Mann-Kendall trends}
\usage{
mk_sig(d_long, alpha = 0.05)
}
\arguments{
\item{d_long}{data frame in long format (see make_long)}

\item{alpha}{significance level (default 0.05)}
}
\value{
table showing number of significant and non-significant Mann-Kendall trends
}
\description{
Count number of significant Mann-Kendall trends
}
\examples{
mk_sig(nine_points_long)
mk_sig(nine_points_long, alpha=0.01)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
