% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rect_hclust}
\alias{rect_hclust}
\title{Dendrogram with clusters shown by rectangles (like rect.hclust)}
\usage{
rect_hclust(object, n_cluster)
}
\arguments{
\item{object}{cluster object from hclust}

\item{n_cluster}{number of clusters to show}
}
\value{
ggplot like output from rect.hclust
}
\description{
Dendrogram with clusters shown by rectangles (like rect.hclust)
}
\examples{
rect_hclust(ward(nine_points, T, 3)$object, 4)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
