% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil-sen.R
\name{theil_sen_map}
\alias{theil_sen_map}
\title{Map of Theil-Sen slopes of time trends with size of slope reflected in size of point}
\usage{
theil_sen_map(d_long, locations, bounding_box, zoom = 5, scaling = 1)
}
\arguments{
\item{d_long}{data frame in long format (see make_long)}

\item{locations}{latitude and longitude of locations}

\item{bounding_box}{bottom left long, lat, top right long, lat (as a vector)}

\item{zoom}{zoom factor for stamen map (default 5)}

\item{scaling}{factor to control overall size of points}
}
\value{
graph
}
\description{
Map of Theil-Sen slopes of time trends with size of slope reflected in size of point
}
\examples{
theil_sen_map(nine_points_long, nine_points_locations, bounding_box=c(-87.5, 71, -82.5, 74), zoom=6, scaling=0.8)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
