% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil-sen.R
\name{theil_sen_slopes}
\alias{theil_sen_slopes}
\title{Theil-Sen slopes of time trends for each location}
\usage{
theil_sen_slopes(d_long)
}
\arguments{
\item{d_long}{data frame in long format (see make_long)}
}
\value{
data frame containing location, nested data frame containing year and days, Theil-Sen slope for that location
}
\description{
Theil-Sen slopes of time trends for each location
}
\examples{
theil_sen_slopes(nine_points_long)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
