% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil-sen.R
\name{theil_sen_summary}
\alias{theil_sen_summary}
\title{Theil-Sen slopes of time trends for each location, summarized}
\usage{
theil_sen_summary(d_long)
}
\arguments{
\item{d_long}{data frame in long format (see make_long)}
}
\value{
numerical summaries of Theil-Sen slopes
}
\description{
Theil-Sen slopes of time trends for each location, summarized
}
\examples{
theil_sen_summary(nine_points_long)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
