% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-trend-plot.R
\name{time_trend_plot}
\alias{time_trend_plot}
\title{Time trend plot, with linear or lowess trend, for all locations}
\usage{
time_trend_plot(d_long, lowess = TRUE)
}
\arguments{
\item{d_long}{data frame in long format (see make_long)}

\item{lowess}{if TRUE, add lowess curve; if FALSE, add linear trend}
}
\value{
ggplot time trend plot with locations in facets
}
\description{
Time trend plot, with linear or lowess trend, for all locations
}
\examples{
time_trend_plot(nine_points_long) # with lowess
time_trend_plot(nine_points_long, FALSE) # with line

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
