% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ward}
\alias{ward}
\title{Hierarchical (Ward) cluster analysis}
\usage{
ward(d, remove_missing = T, n_cluster)
}
\arguments{
\item{d}{data frame in wide format (see make_wide)}

\item{remove_missing}{whether or not to remove missing data}

\item{n_cluster}{number of clusters to obtain}
}
\value{
list: dendrogram, cluster memberships
}
\description{
Hierarchical (Ward) cluster analysis
}
\examples{
ward(nine_points, T, 3)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
