% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-trend-plot.R
\name{time_trend_single}
\alias{time_trend_single}
\title{Time trend plot, with linear or lowess trend, for single location}
\usage{
time_trend_single(d, lowess = TRUE, loc)
}
\arguments{
\item{d}{data frame in wide format (see make_wide)}

\item{lowess}{if TRUE, add lowess curve; if FALSE, add linear trend (defaults to TRUE)}

\item{loc}{character string that identifies location}
}
\value{
ggplot time trend plot for specified location
}
\description{
Time trend plot, with linear or lowess trend, for single location
}
\examples{
time_trend_single(nine_points, loc=6) # with lowess
time_trend_single(nine_points, lowess=FALSE, loc=6) # with line

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
